% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_simultaneous_bands.R
\name{add_simultaneous_bands_sfsmisc}
\alias{add_simultaneous_bands_sfsmisc}
\title{Adds simultaneous confidence band to ECDF using sfsmisc}
\usage{
add_simultaneous_bands_sfsmisc(
  p,
  data,
  response_col,
  group_col,
  ref_level = NULL,
  alpha = 0.05
)
}
\arguments{
\item{p}{\code{ggcdf} ggplot returned by \code{quadruplot()}}

\item{data}{data frame used in \code{quadruplot()}}

\item{response_col}{character giving the name of response variable}

\item{group_col}{character giving the name of group factor}

\item{ref_level}{character giving the reference level of group factor}

\item{alpha}{size of test (0.05) used to provide confidence level}
}
\description{
Adds simultaneous confidence band to ECDF using sfsmisc
}
\keyword{internal}
