% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_simultaneous_bands.R
\name{calc_simultaneous_ecdf_bands_sfsmisc}
\alias{calc_simultaneous_ecdf_bands_sfsmisc}
\title{Confidence bands for ECDF using sfsmisc::KSd}
\usage{
calc_simultaneous_ecdf_bands_sfsmisc(x, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{size of test; hence confidence level is 1 - alpha}
}
\description{
Confidence bands for ECDF using sfsmisc::KSd
}
\keyword{internal}
