% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ex2}
\alias{Ex2}
\title{Synthetic data}
\format{
A data frame with 200 observations on 2 variables.
}
\usage{
data(Ex2)
}
\description{
A data frame with numeric \code{y} and factor \code{group}
}
