% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC plot with confidence band -- internal function}
\usage{
plot_roc(x, ...)
}
\arguments{
\item{x}{Object of class \code{roc_ci} returned by \code{roc_with_ci()}}

\item{...}{not used}
}
\description{
ROC plot with confidence band -- internal function
}
\keyword{internal}
