% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wmw_test.R
\name{print.wmw_test}
\alias{print.wmw_test}
\title{Print Method for wmw_test Objects}
\usage{
\method{print}{wmw_test}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object of class 'wmw_test' returned by \code{wmw_test()}}

\item{digits}{Integer, number of digits to display for numeric results (default: 4)}

\item{...}{Additional arguments (not currently used)}
}
\description{
Prints summary of Wilcoxon-Mann-Whitney discrimination test results.
}
