% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation3}
\alias{simulation3}
\title{Synthetic data}
\format{
A list containing simulation results (N=500, n=300):
\describe{
\item{wmw_ci}{95\% confidence intervals obtained by pseudomedian_ci()}
\item{wt_ci}{95\% confidence intervals obtained by wilcox.test()}
\item{eauc}{Values of eAUC}
\item{pseudomedian}{Values of the pseudomedian}
}
}
\usage{
data(simulation3)
}
\description{
Synthetic data
}
