% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudomedian_ci.R
\name{test_auc_half}
\alias{test_auc_half}
\title{Test H0: AUC = 0.5 (wilcox.test-style edge case handling)}
\usage{
test_auc_half(x, y)
}
\arguments{
\item{x}{first sample}

\item{y}{second sample}
}
\value{
p-value for testing H0: AUC = 0.5
}
\description{
Test for AUC = 0.5 following wilcox.test edge case philosophy
}
