% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shift_equivalence.R
\name{test_shift_equivalence}
\alias{test_shift_equivalence}
\title{Test of equality of distributions from twosamples library applied to median-centered data}
\usage{
test_shift_equivalence(x, y, test = "ks", seed = 123L)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{test}{one of c("ks", "kuiper", "cvm", "ad", "wass", "dts")}

\item{seed}{numeric, used in \code{set.seed()}}
}
\value{
A list of class "shift_test" containing:
\item{statistic}{Test statistic value}
\item{p.value}{P-value of the shift equivalence test}
\item{method}{Character string describing the test method}
\item{alternative}{Character string describing the alternative hypothesis}
\item{data.name}{Character string with the names of the data}
\item{assumptions_met}{Logical indicating if shift equivalence assumptions are satisfied}
}
\description{
Applies a specified test from \code{twosamples} library to median-centered data.
}
\references{
For more details see the \href{https://twosampletest.com/}{Two Sample Test Package Website}

Dowd, C. (2020). A new ECDF two-sample test statistic. arXiv preprint arXiv:2007.01360.
}
\keyword{internal}
