% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation2}
\alias{simulation2}
\title{Synthetic data}
\format{
A list containing simulation results (N=10000, n=1000):
\describe{
\item{eauc}{Empirical AUC values}
\item{pval_wt}{Traditional wilcox.test p-values}
\item{pval_wmw}{WMW p-values under H0: AUC = 0.5}
}
}
\usage{
data(simulation2)
}
\description{
Synthetic data
}
