% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shift_equivalence.R
\name{test_shift_equivalence}
\alias{test_shift_equivalence}
\title{Test of equality of distributions from twosamples library applied to median-centered data}
\usage{
test_shift_equivalence(x, y, test = "ks")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{test}{one of c("ks", "kuiper", "cvm", "ad", "wass", "dts")}
}
\description{
Applies a specified test from \code{twosamples} library to median-centered data.
}
\references{
For more details see the \href{https://twosampletest.com/}{Two Sample Test Package Website}

Dowd, C. (2020). A new ECDF two-sample test statistic. arXiv preprint arXiv:2007.01360.
}
\keyword{internal}
